Script checks for correct distribution
Currently the file checks if you are running CentOS (any version), and if you are not then it quits.  I should probably fix it so that it checks for CentOS 7 to ensure the correct libraries/dependencies are available, but I'm going to leave it as-is for now.  The check also fails if you are running RHEL despite that also being compatible, so that's also on the TODO.