Enabled Python3.6, added sudo to commands
After Python3.6 is installed, it needs to be enabled in order to be usable.  Otherwise any commands using Python 3 won't be recognized, which would completely prevent the Python-based webpage from running properly.  Additionally, sudo was added to the commands that use it.  Previously the script didn't contain any sudo lines, but that required the user to run it as root or with sudo before the script name.  That meant that when the git repo was cloned, its folder was owned by root.  Adding sudo should prevent commands that don't need root from being run under elevated permissions.