corrected an issue with installing git, check for existing folder
At some point during testing I had un-commented the line to install git, so it was no longer available to clone the rest of the repository.  I un-commented the line and removed the rest of the packages it installed, as they were remnants from when I was going to use Python instead of PHP for the web form.  Additionally, I have the script check if the OpenHPC-Setup directory already exists before cloning the repository.