Web page additions
I added multiple fields to the web page, unfortunately it will be very difficult to autofill the majority of them.  I'm doing my best to at least provide input fields for required data in case a manual setup is needed, which will likely be the case.  I'm attempting to add in a javascript section to eliminate sections that are unnecessary, such as disabling unneeded fields for compute nodes.  Currently missing are the network mask, compute node hostnames, name-matching regex, and name prefixes.  If possible, most of this will be automated away in the final product.